+++ tests/unit/mdses/test_holog_mds.py
@@ -8,6 +8,7 @@ from astrohack.utils.verification_tools import (
     add_data_folder_to_names_in_class,
     are_txt_files_equal,
     are_png_files_close,
+    execute_cleanup,
 )
 
 matplotlib.use("Agg")
@@ -38,8 +39,9 @@ class TestHologMDS:
 
     @classmethod
     def teardown_class(cls):
-        shutil.rmtree(cls.data_dir)
-        shutil.rmtree(cls.destination_folder)
+        if execute_cleanup():
+            shutil.rmtree(cls.data_dir)
+            shutil.rmtree(cls.destination_folder)
         return
 
     def test_init(self):
