+++ tests/unit/mdses/test_panel_mds.py
@@ -13,6 +13,7 @@ from astrohack.utils.verification_tools import (
     are_txt_files_equal,
     are_png_files_close,
     are_fits_files_close,
+    execute_cleanup,
 )
 
 matplotlib.use("Agg")
@@ -69,8 +70,9 @@ class TestpanelMDS:
 
     @classmethod
     def teardown_class(cls):
-        shutil.rmtree(cls.data_dir)
-        shutil.rmtree(cls.destination_folder)
+        if execute_cleanup():
+            shutil.rmtree(cls.data_dir)
+            shutil.rmtree(cls.destination_folder)
         return
 
     def test_init(self):
