+++ tests/unit/user_facing_functions/test_holog.py
@@ -10,6 +10,7 @@ from astrohack import open_image
 from astrohack.holog import holog
 from astrohack.utils.verification_tools import (
     add_data_folder_to_names_in_class,
+    execute_cleanup,
 )
 
 
@@ -38,7 +39,8 @@ class TestHolog:
     def teardown_class(cls):
         """teardown any state that was previously setup with a call to setup_class
         such as deleting test data"""
-        shutil.rmtree(cls.data_dir)
+        if execute_cleanup():
+            shutil.rmtree(cls.data_dir)
 
     def test_defaults(self):
         """
