+++ tests/unit/user_facing_functions/test_locit.py
@@ -12,6 +12,7 @@ from astrohack.locit import locit
 from astrohack.utils.verification_tools import (
     are_lists_equal,
     add_data_folder_to_names_in_class,
+    execute_cleanup,
 )
 
 
@@ -47,7 +48,8 @@ class TestLocit:
     def teardown_class(cls):
         """teardown any state that was previously setup with a call to setup_class
         such as deleting test data"""
-        shutil.rmtree(cls.data_dir)
+        if execute_cleanup():
+            shutil.rmtree(cls.data_dir)
 
     def test_defaults(self):
         """
